requirements.txt
flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0
gunicorn==21.2.0
aiohttp==3.8.5 